+++ bench/nested_object_list.py
@@ -10,7 +10,7 @@ from koda_validate import (
     ListValidator,
     StringValidator,
 )
-from koda_validate.typedefs import Valid
+from koda_validate.validated import Valid
 
 
 @dataclass
