+++ bench/one_key_invalid_types.py
@@ -4,7 +4,7 @@ from typing import Any, List
 from pydantic import BaseModel, ValidationError, constr
 
 from koda_validate import DictValidator, StringValidator
-from koda_validate.typedefs import Valid
+from koda_validate.validated import Valid
 
 
 @dataclass
