+++ examples/is_close_predicate.py
@@ -2,7 +2,7 @@ import math
 from dataclasses import dataclass
 
 from koda_validate import FloatValidator, Predicate, Serializable
-from koda_validate.typedefs import Invalid, Valid
+from koda_validate.validated import Invalid, Valid
 
 
 @dataclass
