+++ examples/map_validator.py
@@ -1,5 +1,5 @@
 from koda_validate import IntValidator, MapValidator, StringValidator
-from koda_validate.typedefs import Valid
+from koda_validate.validated import Valid
 
 str_to_int_validator = MapValidator(StringValidator(), IntValidator())
