+++ examples/optional.py
@@ -1,5 +1,5 @@
 from koda_validate import IntValidator, OptionalValidator
-from koda_validate.typedefs import Valid
+from koda_validate.validated import Valid
 
 optional_int_validator = OptionalValidator(IntValidator())
