+++ examples/person_dict_callable.py
@@ -1,5 +1,5 @@
 from koda_validate import *
-from koda_validate.typedefs import Valid
+from koda_validate.validated import Valid
 
 
 def reverse_person_args_tuple(a: str, b: int) -> tuple[int, str]:
