+++ koda_validate/one_of.py
@@ -3,7 +3,8 @@ from typing import Any
 from koda import Either, Either3, First, Second, Third
 from koda._generics import A, B, C
 
-from koda_validate.typedefs import Invalid, Serializable, Valid, Validated, Validator
+from koda_validate.base import Serializable, Validator
+from koda_validate.validated import Invalid, Valid, Validated
 
 
 class OneOf2(Validator[Any, Either[A, B], Serializable]):
