+++ tests/test_integer.py
@@ -11,7 +11,7 @@ from koda_validate import (
     Processor,
     Serializable,
 )
-from koda_validate.typedefs import Invalid, Valid
+from koda_validate.validated import Invalid, Valid
 
 
 class Add1Int(Processor[int]):
