+++ tests/test_lazy.py
@@ -5,7 +5,7 @@ from koda import Just, Maybe, nothing
 
 from koda_validate import IntValidator, Lazy
 from koda_validate.dictionary import DictValidator, KeyNotRequired
-from koda_validate.typedefs import Valid
+from koda_validate.validated import Valid
 
 
 def test_lazy() -> None:
