+++ tests/test_one_of.py
@@ -2,7 +2,7 @@ import pytest
 from koda import First, Second, Third
 
 from koda_validate import FloatValidator, IntValidator, OneOf2, OneOf3, StringValidator
-from koda_validate.typedefs import Invalid, Valid
+from koda_validate.validated import Invalid, Valid
 
 
 def test_one_of2() -> None:
